Distinguish between "auto-update" and "latest" apps
As a follow-up to 40baf90, this adds a separation between apps that
auto-update and those that are marked as latest through new command line
options: `--all` now includes apps that auto-update, while `--force` is
(additionally or exclusively) required for apps marked as latest.

The rationale behind this is to allow upgrades of apps that auto-update
without always reinstalling such marked as latest.