Fix call to deleted method
In commit a3231a03fa171d7e701150957ef4f6f74c2b6948 of the homebrew/brew
repository, the `Cask::Cask#versions` method was removed. We relied
on this method to compute the `formatted_version` we wish to display
to the user once the pinning is done.

To achieve the same result, we can simply call `Cask#installed_version`
instead.